Change the resumable query API (#32)
There were couple of problems with it:

- We don't use the async prefix in async method names in the public
API
- We don't return the result for APIs that returns Success string only

So, I have changed the implementation to fix these problems.

Also, I think it is much better to return the first batch of the results
directly to the user, along with an handle to fetch more or stop.
This is more similar to API we have in JS SDK, makes the possible misuses
impossible (you can't call stop before the start anymore etc.), and makes
more sense for the async version. For the async version, we previously
marked the function async but did not call any async functions, which
was weird.

ALso, I have changed the tests a bit to make them more resillient to
failures due to potentially delayed indexing, which might happen
from time to time.