Add support for new info endpoint fields
We wanted to add embedding model info to the info endpoint, as well
as having some separate fields for dense and sparse indexes.

For now, there is some duplicate information in the top level fields
and the fields in the dense_index, but we might deprecate the top
level fields later on.