Allow upserting data as vector (#41)
* Allow upserting data as vector

I stumbled upon this problem while using the sample code provided
in the console. If the index is created with an embedding model,
we still try to use `Vector` type with only data field. It should
have been `Data`, but I think this is a mistake that can happen
a lot. So, we now allow such usages. If this is wrong, server will
reject it anyway, just like it does.

* bump version to 0.7.1